@@ -8,7 +8,7 @@ use crate::*;
 
 /// Plugin to add a blocker to a registry of blockers.
 ///
-/// E.G: [`UiFocused`] -> CameraControlBlockers to block camera controls on ui hovered.
+/// E.G: `UiFocused` -> CameraControlBlockers to block camera controls on ui hovered.
 #[derive(Default)]
 pub struct BlockerRegistration<Blocker, Registry>
 where
