@@ -16,8 +16,10 @@
 */
 
 use crate::{
-    interaction::{spawn_place_object_2d_workflow, ButtonInputType, PlaceObject2d},
-    site::{CurrentLevel, ModelInstance},
+    interaction::{
+        spawn_place_object_2d_workflow, ButtonInputType, PlaceObject2d, PlaceObjectContinuity,
+    },
+    site::ModelInstance,
 };
 use bevy::{
     app::{App, Plugin},
@@ -27,9 +29,8 @@ use bevy::{
     },
     prelude::*,
 };
-use bevy_impulse::{testing::Resource, Service};
+use crossflow::{testing::Resource, Service};
 use rmf_site_picking::{RunSelector, SelectionNodeResult, SelectorInput};
-use tracing::warn;
 
 #[derive(Default)]
 pub struct ObjectPlacementPlugin {}
@@ -58,18 +59,17 @@ impl ObjectPlacementServices {
 pub struct ObjectPlacement<'w, 's> {
     pub services: Res<'w, ObjectPlacementServices>,
     pub commands: Commands<'w, 's>,
-    current_level: Res<'w, CurrentLevel>,
 }
 
 impl<'w, 's> ObjectPlacement<'w, 's> {
-    pub fn place_object_2d(&mut self, object: ModelInstance<Entity>) {
-        let Some(level) = self.current_level.0 else {
-            warn!("Unble to create [object:?] outside a level");
-            return;
-        };
+    pub fn place_object_2d(
+        &mut self,
+        object: ModelInstance<Entity>,
+        on_placed: PlaceObjectContinuity,
+    ) {
         let state = self
             .commands
-            .spawn(SelectorInput(PlaceObject2d { object, level }))
+            .spawn(SelectorInput(PlaceObject2d { object, on_placed }))
             .id();
         self.send(RunSelector {
             selector: self.services.place_object_2d,
@@ -102,7 +102,7 @@ impl Command for ObjectPlaceCommand {
     fn apply(self, world: &mut World) {
         let mut system_state: SystemState<ObjectPlacement> = SystemState::new(world);
         let mut placement = system_state.get_mut(world);
-        placement.place_object_2d(self.0);
+        placement.place_object_2d(self.0, Default::default());
         system_state.apply(world);
     }
 }
