@@ -15,7 +15,7 @@
  *
 */
 
-use crate::{layers, site::*};
+use crate::{layers::ZLayer, site::*};
 use crate::{CurrentWorkspace, Issue, ValidateWorkspace};
 use bevy::ecs::{hierarchy::ChildOf, relationship::AncestorIter};
 use bevy::pbr::ExtendedMaterial;
@@ -24,6 +24,8 @@ use rmf_site_format::{Edge, LaneMarker};
 use std::collections::{BTreeSet, HashMap};
 use uuid::Uuid;
 
+use bevy::picking::Pickable;
+
 // TODO(MXG): Make this configurable, perhaps even a field in the Lane data
 // so users can customize the lane width per lane.
 pub const LANE_WIDTH: f32 = 0.5;
@@ -31,7 +33,8 @@ pub const DEFAULT_LANE_ARROW_SPEED: f32 = 0.5;
 
 #[derive(Component, Debug, Clone, Copy)]
 pub struct LaneSegments {
-    pub layer: Entity,
+    /// Offset the lane height based on nav graph rank
+    pub rank_layer: Entity,
     pub start: Entity,
     pub mid: Entity,
     pub end: Entity,
@@ -112,7 +115,7 @@ pub fn add_lane_visuals(
             }
         }
 
-        let (lane_material, lane_color, height) = graphs.display_style(associated_graphs);
+        let (lane_material, lane_color, rank) = graphs.display_style(associated_graphs);
         let visibility = if should_display_lane(
             edge,
             associated_graphs,
@@ -133,10 +136,13 @@ pub fn add_lane_visuals(
             .point_in_parent_frame_of(edge.end(), Category::Lane, e)
             .unwrap();
 
-        // Create a "layer" entity that manages the height of the lane,
-        // determined by the DisplayHeight of the graph.
-        let layer = commands
-            .spawn((Transform::from_xyz(0.0, 0.0, height), Visibility::default()))
+        // Create a "layer" entity that manages the height offset of the lane,
+        // determined by its nav graph rank.
+        let rank_layer = commands
+            .spawn((
+                Transform::from_xyz(0.0, 0.0, height_for_rank(rank)),
+                Visibility::default(),
+            ))
             .insert(ChildOf(e))
             .id();
 
@@ -147,8 +153,8 @@ pub fn add_lane_visuals(
                     MeshMaterial3d(lane_material.clone()),
                     lane_tf,
                     Visibility::default(),
+                    ChildOf(rank_layer),
                 ))
-                .insert(ChildOf(layer))
                 .id();
 
             let outline = commands
@@ -157,8 +163,9 @@ pub fn add_lane_visuals(
                     MeshMaterial3d::<StandardMaterial>::default(),
                     Transform::from_translation(-0.000_5 * Vec3::Z),
                     Visibility::Hidden,
+                    ChildOf(mesh),
+                    Pickable::IGNORE,
                 ))
-                .insert(ChildOf(mesh))
                 .id();
 
             (mesh, outline)
@@ -213,7 +220,7 @@ pub fn add_lane_visuals(
                 line_stroke_transform(&start_anchor, &end_anchor, LANE_WIDTH),
                 Visibility::default(),
             ))
-            .insert(ChildOf(layer))
+            .insert(ChildOf(rank_layer))
             .id();
 
         let mid_outline: Entity = commands
@@ -222,26 +229,30 @@ pub fn add_lane_visuals(
                 MeshMaterial3d::<StandardMaterial>::default(),
                 Transform::from_translation(-0.000_5 * Vec3::Z),
                 Visibility::Hidden,
+                ChildOf(mid),
+                Pickable::IGNORE,
             ))
-            .insert(ChildOf(mid))
             .id();
 
-        commands
-            .entity(e)
-            .insert(LaneSegments {
-                layer,
+        let z = ZLayer::Lane.to_z();
+        commands.entity(e).insert((
+            LaneSegments {
+                rank_layer,
                 start,
                 mid,
                 end,
                 outlines: [start_outline, mid_outline, end_outline],
-            })
-            .insert((
-                Transform::from_translation([0., 0., layers::ZLayer::Lane.to_z()].into()),
-                visibility,
-            ))
-            .insert(Category::Lane)
-            .insert(EdgeLabels::StartEnd);
+            },
+            Transform::from_translation([0., 0., z].into()),
+            visibility,
+            Category::Lane,
+            EdgeLabels::StartEnd,
+        ));
     }
+}
+
+fn height_for_rank(rank: f32) -> f32 {
+    rank * (ZLayer::Doormat.to_z() - ZLayer::Lane.to_z())
 }
 
 fn update_lane_visuals(
@@ -360,6 +371,7 @@ pub fn update_changed_lane(
         } else {
             Visibility::Hidden
         };
+
         if *visibility != new_visibility {
             *visibility = new_visibility;
         }
@@ -421,7 +433,17 @@ pub fn update_color_for_lanes(
         (&AssociatedGraphs<Entity>, &LaneSegments),
         (With<LaneMarker>, Changed<AssociatedGraphs<Entity>>),
     >,
-    any_graphs_changed: Query<(), (Changed<DisplayColor>, With<NavGraphMarker>)>,
+    any_graphs_changed: Query<
+        (),
+        (
+            Or<(
+                Changed<DisplayColor>,
+                Changed<RecencyRank<NavGraphMarker>>,
+                Changed<Visibility>,
+            )>,
+            With<NavGraphMarker>,
+        ),
+    >,
     all_lanes: Query<(&AssociatedGraphs<Entity>, &LaneSegments), With<LaneMarker>>,
     graphs: GraphSelect,
     lane_materials: Query<&MeshMaterial3d<ExtendedMaterial<StandardMaterial, LaneArrowMaterial>>>,
@@ -538,6 +560,7 @@ pub fn update_visibility_for_lanes(
             } else {
                 Visibility::Hidden
             };
+
             if *visibility != new_visibility {
                 *visibility = new_visibility;
             }
@@ -557,6 +580,7 @@ pub fn update_visibility_for_lanes(
                 } else {
                     Visibility::Hidden
                 };
+
                 if *visibility != new_visibility {
                     *visibility = new_visibility;
                 }
@@ -566,28 +590,28 @@ pub fn update_visibility_for_lanes(
 
     if graph_change {
         for (_, associated_graphs, segments, _) in &lanes {
-            let (mat, _, height) = graphs.display_style(associated_graphs);
+            let (mat, _, rank) = graphs.display_style(associated_graphs);
             for e in segments.iter() {
                 if let Ok(mut m) = materials.get_mut(e) {
                     *m = MeshMaterial3d(mat.clone());
                 }
             }
 
-            if let Ok(mut tf) = transforms.get_mut(segments.layer) {
-                tf.translation.z = height;
+            if let Ok(mut tf) = transforms.get_mut(segments.rank_layer) {
+                tf.translation.z = height_for_rank(rank);
             }
         }
     } else {
         for (_, associated_graphs, segments) in &lanes_with_changed_association {
-            let (mat, _, height) = graphs.display_style(associated_graphs);
+            let (mat, _, rank) = graphs.display_style(associated_graphs);
             for e in segments.iter() {
                 if let Ok(mut m) = materials.get_mut(e) {
                     *m = MeshMaterial3d(mat.clone());
                 }
             }
 
-            if let Ok(mut tf) = transforms.get_mut(segments.layer) {
-                tf.translation.z = height;
+            if let Ok(mut tf) = transforms.get_mut(segments.rank_layer) {
+                tf.translation.z = height_for_rank(rank);
             }
         }
     }
