@@ -24,12 +24,13 @@ const BILLBOARD_BASE_OFFSET: Vec3 = Vec3::new(0., 0., BILLBOARD_LENGTH / 3. * 0.
 const BILLBOARD_EMPTY_OFFSET: Vec3 = Vec3::new(0., 0., BILLBOARD_LENGTH * 0.5);
 const BILLBOARD_MARGIN: Vec3 = Vec3::new(0., 0., BILLBOARD_LENGTH * 0.9);
 
-#[derive(Component, Clone, Default)]
+#[derive(Component, Clone, Copy, Default)]
 pub struct BillboardMeshes {
     pub base: Option<Entity>,
     pub charging: Option<Entity>,
     pub holding: Option<Entity>,
     pub parking: Option<Entity>,
+    pub mutex_group: Option<Entity>,
     pub empty_billboard: Option<Entity>,
 }
 
@@ -60,25 +61,16 @@ fn should_display_point(
 
 pub fn add_location_visuals(
     mut commands: Commands,
-    locations: Query<
-        (
-            Entity,
-            &Point<Entity>,
-            &AssociatedGraphs<Entity>,
-            &LocationTags,
-        ),
-        Added<LocationTags>,
-    >,
+    locations: Query<(Entity, &Point<Entity>, &AssociatedGraphs<Entity>), Added<LocationTags>>,
     graphs: GraphSelect,
     anchors: AnchorParams,
     child_of: Query<&ChildOf>,
     levels: Query<(), With<LevelElevation>>,
     mut dependents: Query<&mut Dependents, With<Anchor>>,
     assets: Res<SiteAssets>,
-    mut materials: ResMut<Assets<StandardMaterial>>,
     current_level: Res<CurrentLevel>,
 ) {
-    for (e, point, associated_graphs, tags) in &locations {
+    for (e, point, associated_graphs) in &locations {
         if let Ok(mut deps) = dependents.get_mut(point.0) {
             deps.insert(e);
         }
@@ -102,83 +94,6 @@ pub fn add_location_visuals(
             .unwrap()
             + ZLayer::Location.to_z() * Vec3::Z;
 
-        let only_workcell_tags = !tags.iter().any(|t| !t.is_workcell());
-        let mut billboard_meshes = BillboardMeshes::default();
-
-        let base_id = commands.spawn_empty().id();
-        commands.entity(e).add_child(base_id);
-
-        if tags.len() == 0 || only_workcell_tags {
-            // If no location tags, spawn empty billboard marker
-            let new_material = materials
-                .get(&assets.empty_billboard_material)
-                .unwrap()
-                .clone();
-
-            commands.entity(base_id).insert((
-                Mesh3d(assets.billboard_mesh.clone()),
-                MeshMaterial3d(materials.add(new_material)),
-                BillboardMarker {
-                    caption_text: None,
-                    offset: BILLBOARD_EMPTY_OFFSET,
-                    hover_enabled: true,
-                },
-            ));
-            billboard_meshes.empty_billboard = Some(base_id);
-        } else {
-            // If location tags exist, spawn billboard base
-            commands.entity(base_id).insert((
-                Mesh3d(assets.billboard_base_mesh.clone()),
-                MeshMaterial3d(assets.base_billboard_material.clone()),
-                BillboardMarker {
-                    caption_text: None,
-                    offset: BILLBOARD_BASE_OFFSET,
-                    hover_enabled: false,
-                },
-            ));
-            billboard_meshes.base = Some(base_id);
-        }
-
-        let mut offset = BILLBOARD_MARGIN - BILLBOARD_BASE_OFFSET;
-
-        for tag in tags.iter() {
-            let id = commands.spawn_empty().id();
-
-            let (material_handle, text) = match tag {
-                LocationTag::Charger => {
-                    billboard_meshes.charging = Some(id);
-                    (&assets.charger_material, "charging")
-                }
-                LocationTag::ParkingSpot => {
-                    billboard_meshes.parking = Some(id);
-                    (&assets.parking_material, "parking")
-                }
-                LocationTag::HoldingPoint => {
-                    billboard_meshes.holding = Some(id);
-                    (&assets.holding_point_material, "holding")
-                }
-                // Workcells are not visualized
-                LocationTag::Workcell(_) => continue,
-            };
-
-            // New material instance created for each billboard as the AlphaMode of each billboard is toggled on hover
-            let new_material = materials.get(material_handle).unwrap().clone();
-
-            commands.entity(id).insert((
-                Mesh3d(assets.billboard_mesh.clone()),
-                MeshMaterial3d(materials.add(new_material)),
-                BillboardMarker {
-                    caption_text: Some(text.to_string()),
-                    offset: offset,
-                    hover_enabled: true,
-                },
-            ));
-
-            commands.entity(e).add_child(id);
-            offset += BILLBOARD_MARGIN;
-        }
-
-        // TODO(MXG): Put icons on the different visual squares based on the location tags
         commands
             .entity(e)
             .insert((
@@ -188,7 +103,6 @@ pub fn add_location_visuals(
                 visibility,
             ))
             .insert(Category::Location)
-            .insert(billboard_meshes)
             .insert(VisualCue::outline());
     }
 }
@@ -266,66 +180,81 @@ pub fn update_location_for_changed_location_tags(
         (
             Entity,
             &LocationTags,
-            &mut BillboardMeshes,
-            &mut Hovered,
-            &mut Selected,
+            &Affiliation<Entity>,
+            Option<&BillboardMeshes>,
+            Option<&mut Hovered>,
+            Option<&mut Selected>,
         ),
-        Changed<LocationTags>,
+        Or<(Changed<LocationTags>, Changed<Affiliation<Entity>>)>,
     >,
     mut billboards: Query<&mut BillboardMarker, With<BillboardMarker>>,
+    mutex_groups: Query<&NameInSite, (With<MutexMarker>, With<Group>)>,
     assets: Res<SiteAssets>,
     mut materials: ResMut<Assets<StandardMaterial>>,
 ) {
-    for (e, tags, mut billboard_meshes, mut hovered, mut selected) in &mut locations {
+    for (e, tags, mutex_group, previous_billboard_meshes, mut hovered, mut selected) in
+        &mut locations
+    {
+        let mut billboard_meshes = previous_billboard_meshes.copied().unwrap_or_default();
         select.write(Select::new(Some(e)));
 
         let only_workcell_tags = !tags.iter().any(|t| !t.is_workcell());
+        let no_billboards = only_workcell_tags && mutex_group.0.is_none();
+
+        let mut remove_interactions = |id| {
+            if let Some(hovered) = &mut hovered {
+                hovered.support_hovering.remove(&id);
+            }
+            if let Some(selected) = &mut selected {
+                selected.support_selected.remove(&id);
+            }
+        };
 
         // Despawn unused billboards
         if let Some(id) = billboard_meshes.empty_billboard {
-            if tags.len() > 0 || !only_workcell_tags {
+            if !no_billboards {
                 commands.entity(id).despawn();
                 billboard_meshes.empty_billboard = None;
-
-                hovered.support_hovering.remove(&id);
-                selected.support_selected.remove(&id);
+                remove_interactions(id);
             }
         }
         if let Some(id) = billboard_meshes.charging {
             if !tags.iter().any(|t| t.is_charger()) {
                 commands.entity(id).despawn();
                 billboard_meshes.charging = None;
-
-                hovered.support_hovering.remove(&id);
-                selected.support_selected.remove(&id);
+                remove_interactions(id);
             }
         }
         if let Some(id) = billboard_meshes.holding {
             if !tags.iter().any(|t| t.is_holding_point()) {
                 commands.entity(id).despawn();
                 billboard_meshes.holding = None;
-
-                hovered.support_hovering.remove(&id);
-                selected.support_selected.remove(&id);
+                remove_interactions(id);
             }
         }
         if let Some(id) = billboard_meshes.parking {
             if !tags.iter().any(|t| t.is_parking_spot()) {
                 commands.entity(id).despawn();
                 billboard_meshes.parking = None;
-
-                hovered.support_hovering.remove(&id);
-                selected.support_selected.remove(&id);
+                remove_interactions(id);
             }
         }
-        if tags.len() == 0 || only_workcell_tags {
+        if let Some(id) = billboard_meshes.mutex_group {
+            if mutex_group.0.is_none() {
+                commands.entity(id).despawn();
+                billboard_meshes.mutex_group = None;
+                remove_interactions(id);
+            }
+        }
+
+        if no_billboards {
             if let Some(id) = billboard_meshes.base {
                 commands.entity(id).despawn();
                 billboard_meshes.base = None;
             }
         }
 
-        if (tags.len() == 0 || only_workcell_tags) && billboard_meshes.empty_billboard.is_none() {
+        if no_billboards && billboard_meshes.empty_billboard.is_none() {
             // If no location tags exist and no empty billboard marker spawned, spawn empty billboard marker
             let id = commands.spawn_empty().id();
             let new_material = materials
@@ -373,17 +302,17 @@ pub fn update_location_for_changed_location_tags(
             };
 
             // If there exists a spawned billboard for this tag, shift existing billboard
-            if existing_billboard_id.is_some() {
-                let Some(billboard_id) = existing_billboard_id else {
-                    return;
-                };
-                let Ok(mut marker) = billboards.get_mut(billboard_id) else {
-                    warn!("Billboard not found");
-                    return;
-                };
-                marker.offset = offset;
-                offset += BILLBOARD_MARGIN;
-                continue;
+            if let Some(billboard_id) = existing_billboard_id {
+                if let Ok(mut marker) = billboards.get_mut(billboard_id) {
+                    marker.offset = offset;
+                    offset += BILLBOARD_MARGIN;
+
+                    continue;
+                }
+
+                error!(
+                    "Invalid billboard entity [{billboard_id:?}]. Overriding with a new billboard entity."
+                );
             }
 
             // There is no existing billboard for this tag, hence spawn new billboard
@@ -392,28 +321,29 @@ pub fn update_location_for_changed_location_tags(
             let (material_handle, text) = match tag {
                 LocationTag::Charger => {
                     billboard_meshes.charging = Some(id);
-                    (&assets.charger_material, "charging")
+                    (&assets.charger_material, "charging".to_string())
                 }
                 LocationTag::ParkingSpot => {
                     billboard_meshes.parking = Some(id);
-                    (&assets.parking_material, "parking")
+                    (&assets.parking_material, "parking".to_string())
                 }
                 LocationTag::HoldingPoint => {
                     billboard_meshes.holding = Some(id);
-                    (&assets.holding_point_material, "holding")
+                    (&assets.holding_point_material, "holding".to_string())
                 }
                 // Workcells are not visualized
                 LocationTag::Workcell(_) => continue,
             };
 
-            // New material instance created for each billboard as the AlphaMode of each billboard is toggled on hover
             let new_material = materials.get(material_handle).unwrap().clone();
 
             commands.entity(id).insert((
                 Mesh3d(assets.billboard_mesh.clone()),
+                // A separate copy of the material is created for each billboard
+                // because we adjust their alpha properties during interaction.
                 MeshMaterial3d(materials.add(new_material)),
                 BillboardMarker {
-                    caption_text: Some(text.to_string()),
+                    caption_text: Some(text),
                     offset: offset,
                     hover_enabled: true,
                 },
@@ -422,6 +352,53 @@ pub fn update_location_for_changed_location_tags(
             commands.entity(e).add_child(id);
             offset += BILLBOARD_MARGIN;
         }
+
+        if let Some(mutex_group) = mutex_group.0 {
+            let mutex_group_text = if let Ok(name) = mutex_groups.get(mutex_group) {
+                format!("mutex group: {}", name.0)
+            } else {
+                String::from("<invalid mutex group>")
+            };
+
+            let mut make_new_billboard = true;
+            if let Some(existing_billboard_id) = billboard_meshes.mutex_group {
+                if let Ok(mut marker) = billboards.get_mut(existing_billboard_id) {
+                    marker.offset = offset;
+                    offset += BILLBOARD_MARGIN;
+
+                    marker.caption_text = Some(mutex_group_text.clone());
+                    make_new_billboard = false;
+                } else {
+                    error!(
+                        "Invalid billboard entity [{existing_billboard_id:?}]. Overriding with a new billboard entity."
+                    );
+                }
+            }
+
+            if make_new_billboard {
+                let material = materials.get(&assets.lockpad_material).unwrap().clone();
+                let id = commands
+                    .spawn((
+                        Mesh3d(assets.billboard_mesh.clone()),
+                        // A separate copy of the material is created for each billboard
+                        // because we adjust their alpha properties during interaction.
+                        MeshMaterial3d(materials.add(material)),
+                        BillboardMarker {
+                            caption_text: Some(mutex_group_text),
+                            offset: offset,
+                            hover_enabled: true,
+                        },
+                        ChildOf(e),
+                    ))
+                    .id();
+
+                billboard_meshes.mutex_group = Some(id);
+
+                offset += BILLBOARD_MARGIN;
+            }
+        }
+
+        commands.entity(e).insert(billboard_meshes);
     }
 }
 
