@@ -38,8 +38,8 @@ use bevy::{
     render::{mesh::VertexAttributeValues, view::RenderLayers},
     scene::SceneInstance,
 };
-use bevy_impulse::*;
 use bevy_mod_outline::{GenerateOutlineNormalsSettings, OutlineMeshExt};
+use crossflow::prelude::*;
 use rmf_site_camera::MODEL_PREVIEW_LAYER;
 use rmf_site_format::{
     Affiliation, AssetSource, Group, Inclusion, IssueKey, ModelInstance, ModelMarker,
@@ -66,7 +66,7 @@ pub struct PendingModel;
 /// For a given `AssetSource`, return all the sources that we should try loading.
 pub fn get_all_for_source(source: &AssetSource) -> Vec<AssetSource> {
     match source {
-        AssetSource::Search(ref name) => {
+        AssetSource::Search(name) => {
             let split: SmallVec<[&str; 8]> = name.split('/').collect();
             let model_name = split.last().unwrap();
             let mut paths = common_model_directory_layouts(&name, model_name);
@@ -157,7 +157,7 @@ fn load_asset_source(
     asset_server: Res<AssetServer>,
     current_workspace: Option<Res<CurrentWorkspace>>,
     site_files: Query<&DefaultFile>,
-) -> impl Future<Output = Result<UntypedHandle, ModelLoadingErrorKind>> {
+) -> impl Future<Output = Result<UntypedHandle, ModelLoadingErrorKind>> + use<> {
     let asset_server = asset_server.clone();
     let base_path = current_workspace.and_then(|w| get_current_workspace_path(w, site_files));
 
@@ -173,7 +173,7 @@ fn load_asset_source(
                     return Err(ModelLoadingErrorKind::InvalidAssetSource(err.to_string()));
                 }
             };
-            asset_server
+            let handle = asset_server
                 .load_untyped_async(&asset_path)
                 .await
                 .map_err(|err| {
@@ -187,7 +187,15 @@ fn load_asset_source(
                         error!("Failed attempt to load asset with [{asset_path}]: {err}");
                     }
                     ModelLoadingErrorKind::AssetServerError(err.to_string())
+                })?;
+            asset_server
+                .wait_for_asset_untyped(&handle)
+                .await
+                .map_err(|err| {
+                    error!("Failed attempt to load asset with [{asset_path}]: {err}");
+                    ModelLoadingErrorKind::AssetServerError(err.to_string())
                 })
+                .map(|_| handle)
         }
     }
 }
@@ -307,7 +315,7 @@ fn handle_model_loading(
         request, channel, ..
     }): AsyncServiceInput<ModelLoadingRequest>,
     model_services: Res<ModelLoadingServices>,
-) -> impl Future<Output = Result<ModelLoadingRequest, ModelLoadingError>> {
+) -> impl Future<Output = Result<ModelLoadingRequest, ModelLoadingError>> + use<> {
     let check_scene_is_spawned = model_services.check_scene_is_spawned.clone();
     async move {
         let sources = get_all_for_source(&request.source);
@@ -394,9 +402,9 @@ fn handle_model_loading_errors(
     mut scene_spawner: ResMut<SceneSpawner>,
     mut delete: EventWriter<Delete>,
 ) -> ModelLoadingResult {
-    let parent = match result {
-        Ok(ref success) => success.request.parent,
-        Err(ref err) => {
+    let parent = match &result {
+        Ok(success) => success.request.parent,
+        Err(err) => {
             let parent = err.request.parent;
             // There was an actual error, cleanup the scene
             if let Ok(scene) = model_scenes.get(parent) {
@@ -479,7 +487,7 @@ impl<'w, 's> ModelLoader<'w, 's> {
         &mut self,
         parent: Entity,
         instance: ModelInstance<Entity>,
-        impulse: impl FnOnce(Impulse<InstanceSpawningResult, ()>),
+        impulse: impl FnOnce(Series<InstanceSpawningResult, ()>),
     ) -> EntityCommands<'_> {
         let affiliation = instance.description.clone();
         let id = self
@@ -504,19 +512,19 @@ impl<'w, 's> ModelLoader<'w, 's> {
     /// Run a basic workflow to update the asset source of an existing entity
     pub fn update_asset_source(&mut self, entity: Entity, source: AssetSource) {
         let interaction = DragPlaneBundle::new(entity, Vec3::Z);
-        self.update_asset_source_impulse(entity, source, Some(interaction))
+        self.update_asset_source_series(entity, source, Some(interaction))
             .detach();
     }
 
     /// Update an asset source and then keep attaching impulses to its outcome.
     /// Remember to call `.detach()` when finished or else the whole chain will be
     /// dropped right away.
-    pub fn update_asset_source_impulse(
+    pub fn update_asset_source_series(
         &mut self,
         entity: Entity,
         source: AssetSource,
         interaction: Option<DragPlaneBundle>,
-    ) -> Impulse<'w, 's, '_, ModelLoadingResult, ()> {
+    ) -> Series<'w, 's, '_, ModelLoadingResult, ()> {
         self.commands.request(
             ModelLoadingRequest::new(entity, source, interaction),
             self.services
@@ -539,7 +547,7 @@ impl<'w, 's> ModelLoader<'w, 's> {
         }
         let interaction = DragPlaneBundle::new(entity, Vec3::Z);
         for e in instance_entities.iter() {
-            self.update_asset_source_impulse(*e, source.clone(), Some(interaction.clone()))
+            self.update_asset_source_series(*e, source.clone(), Some(interaction.clone()))
                 .detach();
         }
     }
@@ -552,7 +560,7 @@ fn load_model_dependencies(
     children_q: Query<&Children>,
     models: Query<&AssetSource, With<ModelMarker>>,
     model_loading: Res<ModelLoadingServices>,
-) -> impl Future<Output = Result<ModelLoadingRequest, ModelLoadingError>> {
+) -> impl Future<Output = Result<ModelLoadingRequest, ModelLoadingError>> + use<> {
     let models = DescendantIter::new(&children_q, request.parent)
         .filter_map(|c| models.get(c).ok().map(|source| (c, source.clone())))
         .collect::<Vec<_>>();
@@ -603,7 +611,7 @@ impl ModelLoadingServices {
         )
         .add_systems(
             PostUpdate,
-            (ApplyDeferred, flush_impulses()).in_set(ModelLoadingSet::CheckSceneFlush),
+            (ApplyDeferred, flush_execution()).in_set(ModelLoadingSet::CheckSceneFlush),
         );
         let check_scene_is_spawned = app.spawn_continuous_service(
             PostUpdate,
