@@ -15,7 +15,7 @@
  *
 */
 
-use crate::{layers::ZLayer, site::*};
+use crate::site::*;
 use bevy::{ecs::system::SystemParam, prelude::*};
 
 #[derive(SystemParam)]
@@ -82,18 +82,11 @@ impl<'w, 's> GraphSelect<'w, 's> {
                 .max_by(|(_, _, _, _, a), (_, _, _, _, b)| a.cmp(b))
                 .map(|(_, m, c, _, d)| (m.clone(), *c, *d)),
         }
-        .map(|(m, c, d)| {
-            (
-                m.0,
-                c.to_bevy(),
-                d.proportion() * (ZLayer::Doormat.to_z() - ZLayer::Lane.to_z())
-                    + ZLayer::Lane.to_z(),
-            )
-        })
+        .map(|(m, c, d)| (m.0, c.to_bevy(), d.proportion()))
         .unwrap_or((
             self.assets.unassigned_lane_material.clone(),
             NAV_UNASSIGNED_COLOR,
-            ZLayer::Doormat.to_z(),
+            0.99,
         ))
     }
 
