@@ -20,9 +20,9 @@ use crate::site::{
     MechanicalSystem, Mobility, ModelMarker, ModelProperty, ModelPropertyQuery, PowerSource, Robot,
     RobotProperty, RobotPropertyKind,
 };
-use bevy::prelude::*;
+use bevy::{ecs::relationship::AncestorIter, prelude::*};
 use rmf_site_format::robot_properties::*;
-use sdformat_rs::{ElementData, ElementMap, XmlElement};
+use sdformat::{ElementData, ElementMap, XmlElement};
 use serde_json::{Map, Value};
 
 pub struct SlotcarSdfPlugin;
@@ -43,10 +43,15 @@ impl Plugin for SlotcarSdfPlugin {
     }
 }
 
+/// When loading SDF models, some RobotPropertyKinds may be directly inserted into
+/// a model instance descendant instead of the instance or affiliated description.
+/// This system checks for such cases and inserts the respective components into
+/// the affiliated model description (if there is no other RobotProperty data
+/// present) and remove it from the original entity.
 fn insert_slotcar_components(
     mut commands: Commands,
-    is_static: Query<&ModelProperty<IsStatic>, (With<ModelMarker>, With<Group>)>,
-    robot_property_kinds: Query<
+    slotcar_property_kinds: Query<
+        // All 4 components will be present (see sdf_loader.rs)
         (
             Entity,
             &DifferentialDrive,
@@ -56,43 +61,45 @@ fn insert_slotcar_components(
         ),
         (Without<ModelMarker>, Without<Group>),
     >,
-    robot_properties: Query<(&Mobility, &PowerSource), (With<ModelMarker>, With<Group>)>,
-    model_descriptions: Query<&ModelProperty<Robot>, (With<ModelMarker>, With<Group>)>,
+    model_descriptions: Query<
+        (
+            Option<&ModelProperty<Robot>>,
+            Option<&ModelProperty<IsStatic>>,
+        ),
+        (With<ModelMarker>, With<Group>),
+    >,
     model_instances: ModelPropertyQuery<Robot>,
     child_of: Query<&ChildOf>,
 ) {
     for (e, differential_drive, battery, ambient_system, mechanical_system) in
-        robot_property_kinds.iter()
+        slotcar_property_kinds.iter()
     {
-        if !model_descriptions.get(e).is_ok() {
-            // A non-description entity has the RobotPropertyKind component, it could have been inserted into a
-            // model instance descendent when processing importing robot plugins
-            // Insert this component in the affiliated description and remove it from the original entity
-            let mut description_entity: Option<Entity> = None;
-            let mut target_entity: Entity = e;
-            while let Ok(parent) = child_of.get(target_entity).map(|co| co.parent()) {
-                if let Some(desc) = model_instances.get(parent).ok().and_then(|a| a.0) {
-                    if !robot_properties.get(desc).is_ok()
-                        && is_static.get(desc).is_ok_and(|is| !is.0 .0)
-                    {
-                        description_entity = Some(desc);
-                    }
-                    break;
+        let description_entity = AncestorIter::new(&child_of, e)
+            .find(|p| model_instances.get(*p).is_ok())
+            .and_then(|parent| model_instances.get(parent).ok().and_then(|a| a.0));
+
+        if let Some(((opt_desc_robot, opt_desc_is_static), desc)) =
+            description_entity.and_then(|e| model_descriptions.get(e).ok().zip(Some(e)))
+        {
+            let mut robot = Robot::default();
+
+            if let Some(desc_robot) = opt_desc_robot {
+                if !desc_robot.0.properties.is_empty() {
+                    // The robot properties for this model description has already
+                    // been populated, do not overwrite
+                    commands
+                        .entity(e)
+                        .remove::<DifferentialDrive>()
+                        .remove::<Battery>()
+                        .remove::<AmbientSystem>()
+                        .remove::<MechanicalSystem>();
+                    continue;
                 }
-                target_entity = parent;
+                robot = desc_robot.0.clone();
             }
 
-            if let Some(desc) = description_entity {
-                let mut robot = match model_descriptions.get(desc) {
-                    Ok(ModelProperty(r)) => r.clone(),
-                    Err(_) => Robot::default(),
-                };
-
-                // NOTE(@xiyuoh) It's probably a better idea to generalize this system
-                // for each RobotPropertyKind, but the current design of ChangePlugin
-                // and Robot will result in the later system run to overwrite changes
-                // from the earlier system. For now we will process data from Battery/
-                // DifferentialDrive/MechanicalSystem/AmbientSystem in a single system.
+            // Only insert Mobility if robot is not static
+            if !opt_desc_is_static.is_some_and(|is_static| is_static.0 .0) {
                 if let Ok(mobility_value) = serialize_robot_property_from_kind::<
                     Mobility,
                     DifferentialDrive,
@@ -100,49 +107,45 @@ fn insert_slotcar_components(
                 {
                     robot.properties.insert(Mobility::label(), mobility_value);
                 }
-
-                if let Ok(power_source_value) =
-                    serialize_robot_property_from_kind::<PowerSource, Battery>(battery.clone())
-                {
-                    robot
-                        .properties
-                        .insert(PowerSource::label(), power_source_value);
-                }
-
-                let mut power_dissipation_config = Map::new();
-                if let Some(power_dissipation_map) = power_dissipation_config
-                    .entry("config")
-                    .or_insert(Value::Object(Map::new()))
-                    .as_object_mut()
-                {
-                    if let Ok(mechanical_system_value) =
-                        serialize_robot_property_kind::<MechanicalSystem>(mechanical_system.clone())
-                    {
-                        power_dissipation_map
-                            .insert(MechanicalSystem::label(), mechanical_system_value);
-                    }
-                    if let Ok(ambient_system_value) =
-                        serialize_robot_property_kind::<AmbientSystem>(ambient_system.clone())
-                    {
-                        power_dissipation_map.insert(AmbientSystem::label(), ambient_system_value);
-                    }
-                }
-                if !power_dissipation_config.is_empty() {
-                    robot.properties.insert(
-                        PowerDissipation::label(),
-                        Value::Object(power_dissipation_config),
-                    );
-                }
-
-                commands.trigger(Change::new(ModelProperty(robot), desc));
             }
-            commands
-                .entity(e)
-                .remove::<DifferentialDrive>()
-                .remove::<Battery>()
-                .remove::<AmbientSystem>()
-                .remove::<MechanicalSystem>();
+
+            if let Ok(power_source_value) =
+                serialize_robot_property_from_kind::<PowerSource, Battery>(battery.clone())
+            {
+                robot
+                    .properties
+                    .insert(PowerSource::label(), power_source_value);
+            }
+
+            let mut power_dissipation_map = Map::new();
+            if let Ok(mechanical_system_value) =
+                serialize_robot_property_kind::<MechanicalSystem>(mechanical_system.clone())
+            {
+                power_dissipation_map.insert(MechanicalSystem::label(), mechanical_system_value);
+            }
+            if let Ok(ambient_system_value) =
+                serialize_robot_property_kind::<AmbientSystem>(ambient_system.clone())
+            {
+                power_dissipation_map.insert(AmbientSystem::label(), ambient_system_value);
+            }
+            if !power_dissipation_map.is_empty() {
+                let mut power_dissipation_config = Map::new();
+                power_dissipation_config
+                    .insert("config".to_string(), Value::Object(power_dissipation_map));
+                robot.properties.insert(
+                    PowerDissipation::label(),
+                    Value::Object(power_dissipation_config),
+                );
+            }
+
+            commands.trigger(Change::new(ModelProperty(robot), desc).or_insert());
         }
+        commands
+            .entity(e)
+            .remove::<DifferentialDrive>()
+            .remove::<Battery>()
+            .remove::<AmbientSystem>()
+            .remove::<MechanicalSystem>();
     }
 }
 
@@ -400,7 +403,7 @@ impl SlotcarParams {
     }
 }
 
-fn slotcar_export_handler(In(input): In<(Entity, Value)>) -> sdformat_rs::XmlElement {
+fn slotcar_export_handler(In(input): In<(Entity, Value)>) -> sdformat::XmlElement {
     let (_, slotcar_config) = input;
     let mut slotcar_params = SlotcarParams::default();
 
