@@ -25,7 +25,9 @@ pub enum PortingError {
     InvalidLevelName(String),
     #[error("a non-existent lift {0} was referenced")]
     InvalidLiftName(String),
-    #[error("wrong number [{door_count}] of lift cabin doors for lift [{lift}]; must be no greater than 4")]
+    #[error(
+        "wrong number [{door_count}] of lift cabin doors for lift [{lift}]; must be no greater than 4"
+    )]
     InvalidLiftCabinDoorCount { lift: String, door_count: usize },
     #[error("lift [{lift}] has more than one cabin door on {face} side")]
     DuplicateLiftCabinDoor { lift: String, face: &'static str },
