@@ -16,10 +16,13 @@
 */
 
 use crate::*;
-#[cfg(feature = "bevy")]
-use bevy::prelude::{Component, Reflect};
-use serde::{de::DeserializeOwned, Deserialize, Serialize};
+use serde::{Deserialize, Serialize};
 use std::fmt;
+#[cfg(feature = "bevy")]
+use {
+    bevy::prelude::{Component, Reflect},
+    serde::de::DeserializeOwned,
+};
 
 #[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
 #[cfg_attr(feature = "bevy", derive(Component))]
