@@ -17,8 +17,8 @@
 
 use bevy_app::{App, Last, Plugin};
 use bevy_ecs::prelude::*;
-use bevy_impulse::*;
 use bevy_input::prelude::*;
+use crossflow::*;
 
 pub enum ButtonInputType {
     Pressed,
@@ -57,21 +57,17 @@ fn keyboard_pressed_stream(
         orders.for_each(|order| {
             order
                 .streams()
-                .send(StreamOf((*key_code, ButtonInputType::JustReleased)))
+                .send((*key_code, ButtonInputType::JustReleased))
         });
     }
     for key_code in keyboard_input.get_pressed() {
-        orders.for_each(|order| {
-            order
-                .streams()
-                .send(StreamOf((*key_code, ButtonInputType::Pressed)))
-        });
+        orders.for_each(|order| order.streams().send((*key_code, ButtonInputType::Pressed)));
     }
     for key_code in keyboard_input.get_just_pressed() {
         orders.for_each(|order| {
             order
                 .streams()
-                .send(StreamOf((*key_code, ButtonInputType::JustPressed)))
+                .send((*key_code, ButtonInputType::JustPressed))
         });
     }
 }
