@@ -1,11 +1,11 @@
 use std::marker::PhantomData;
 
 use bevy_ecs::system::ScheduleSystem;
-use bevy_impulse::{
-    AddContinuousServicesExt, AddServicesExt, IntoBlockingCallback, QuickContinuousServiceBuild,
-    RequestExt, RunCommandsOnWorldExt, ScheduleConfigs, Service, SpawnWorkflowExt, flush_impulses,
-};
 use bytemuck::TransparentWrapper;
+use crossflow::{
+    AddContinuousServicesExt, AddServicesExt, IntoBlockingCallback, QuickContinuousServiceBuild,
+    RequestExt, RunCommandsOnWorldExt, ScheduleConfigs, Service, SpawnWorkflowExt, flush_execution,
+};
 use std::fmt::Debug;
 
 use bevy_app::prelude::*;
@@ -17,9 +17,9 @@ type SelectionService = Service<(), ()>;
 
 /// Plugin for default selection behaviour within rmf_site.
 ///
-/// [`T`] is the default service for selection behaviour within this plugin.
+/// `T` is the default service for selection behaviour within this plugin.
 ///
-/// !!! Ensure [`T`]'s assocaited plugin is initialized before this one or this plugin will crash !!!
+/// !!! Ensure `T`'s assocaited plugin is initialized before this one or this plugin will crash !!!
 pub struct SelectionPlugin<T>
 where
     T: Debug + Send + Sync + Resource + TransparentWrapper<SelectionService> + 'static,
@@ -69,13 +69,13 @@ where
         .add_systems(
             Update,
             (
-                (ApplyDeferred, flush_impulses())
+                (ApplyDeferred, flush_execution())
                     .chain()
                     .in_set(SelectionServiceStages::PickFlush),
-                (ApplyDeferred, flush_impulses())
+                (ApplyDeferred, flush_execution())
                     .chain()
                     .in_set(SelectionServiceStages::HoverFlush),
-                (ApplyDeferred, flush_impulses())
+                (ApplyDeferred, flush_execution())
                     .chain()
                     .in_set(SelectionServiceStages::SelectFlush),
             ),
@@ -134,7 +134,6 @@ impl Plugin for InspectorServicePlugin {
                 .then_node(keyboard_pressed)
                 .streams
                 .chain(builder)
-                .inner()
                 .then(deselect_on_esc.into_blocking_callback())
                 .unused();
             let selection = fork_input
@@ -142,7 +141,7 @@ impl Plugin for InspectorServicePlugin {
                 .then_node(inspector_select_service);
             selection
                 .streams
-                .1
+                .select
                 .chain(builder)
                 .then(selection_update)
                 .unused();
